chore: Update app launcher icons
This commit adds new launcher icons for the app. The ic_launcher_background.xml file is added to the drawable folder, and the ic_launcher.xml and ic_launcher_round.xml files are added to the mipmap-anydpi-v26 folder. These icons are used to represent the app on the device's home screen and app drawer.